Refactor utility methods: parsePeriodString(), createPeriodString(), getRealClientIpAddr() and displayBytes()
