+++ az-hdfs-viewer/src/main/resources/azkaban/viewer/hdfs/velocity/hdfs-file.vm
@@ -96,9 +96,10 @@
                 </tr>
                 <tr>
                   <td class="property-key">Size</td>
-                  <td class="property-value-half">$utils.displayBytes(${status.len})</td>
+                  <td class="property-value-half">$WebUtils.displayBytes(${status.len})</td>
                   <td class="property-key">Block Size</td>
-                  <td class="property-value-half">$utils.displayBytes(${status.getBlockSize()})</td>
+                  <td class="property-value-half">$WebUtils.displayBytes(${status.getBlockSize()})
+                  </td>
                 </tr>
                 <tr>
                   <td class="property-key">Reps</td>
