+++ azkaban-common/src/main/java/azkaban/scheduler/Schedule.java
@@ -19,6 +19,7 @@ package azkaban.scheduler;
 import azkaban.executor.ExecutionOptions;
 import azkaban.sla.SlaOption;
 import azkaban.utils.Pair;
+import azkaban.utils.TimeUtils;
 import azkaban.utils.Utils;
 import java.util.ArrayList;
 import java.util.Date;
@@ -27,14 +28,7 @@ import java.util.List;
 import java.util.Map;
 import org.joda.time.DateTime;
 import org.joda.time.DateTimeZone;
-import org.joda.time.Days;
-import org.joda.time.DurationFieldType;
-import org.joda.time.Hours;
-import org.joda.time.Minutes;
-import org.joda.time.Months;
 import org.joda.time.ReadablePeriod;
-import org.joda.time.Seconds;
-import org.joda.time.Weeks;
 import org.quartz.CronExpression;
 
 public class Schedule {
@@ -91,72 +85,6 @@ public class Schedule {
     this.cronExpression = cronExpression;
   }
 
-  public static ReadablePeriod parsePeriodString(final String periodStr) {
-    final ReadablePeriod period;
-    final char periodUnit = periodStr.charAt(periodStr.length() - 1);
-    if (periodUnit == 'n') {
-      return null;
-    }
-
-    final int periodInt =
-        Integer.parseInt(periodStr.substring(0, periodStr.length() - 1));
-    switch (periodUnit) {
-      case 'M':
-        period = Months.months(periodInt);
-        break;
-      case 'w':
-        period = Weeks.weeks(periodInt);
-        break;
-      case 'd':
-        period = Days.days(periodInt);
-        break;
-      case 'h':
-        period = Hours.hours(periodInt);
-        break;
-      case 'm':
-        period = Minutes.minutes(periodInt);
-        break;
-      case 's':
-        period = Seconds.seconds(periodInt);
-        break;
-      default:
-        throw new IllegalArgumentException("Invalid schedule period unit '"
-            + periodUnit);
-    }
-
-    return period;
-  }
-
-  public static String createPeriodString(final ReadablePeriod period) {
-    String periodStr = "n";
-
-    if (period == null) {
-      return "n";
-    }
-
-    if (period.get(DurationFieldType.months()) > 0) {
-      final int months = period.get(DurationFieldType.months());
-      periodStr = months + "M";
-    } else if (period.get(DurationFieldType.weeks()) > 0) {
-      final int weeks = period.get(DurationFieldType.weeks());
-      periodStr = weeks + "w";
-    } else if (period.get(DurationFieldType.days()) > 0) {
-      final int days = period.get(DurationFieldType.days());
-      periodStr = days + "d";
-    } else if (period.get(DurationFieldType.hours()) > 0) {
-      final int hours = period.get(DurationFieldType.hours());
-      periodStr = hours + "h";
-    } else if (period.get(DurationFieldType.minutes()) > 0) {
-      final int minutes = period.get(DurationFieldType.minutes());
-      periodStr = minutes + "m";
-    } else if (period.get(DurationFieldType.seconds()) > 0) {
-      final int seconds = period.get(DurationFieldType.seconds());
-      periodStr = seconds + "s";
-    }
-
-    return periodStr;
-  }
-
   public ExecutionOptions getExecutionOptions() {
     return this.executionOptions;
   }
@@ -189,7 +117,7 @@ public class Schedule {
     } else if (this.cronExpression != null) {
       return underlying + " with CronExpression {" + this.cronExpression + "}";
     } else {
-      return underlying + " with precurring period of " + createPeriodString(this.period);
+      return underlying + " with precurring period of " + TimeUtils.createPeriodString(this.period);
     }
   }
