+++ azkaban-common/src/main/java/azkaban/trigger/builtin/BasicTimeChecker.java
@@ -17,6 +17,7 @@
 package azkaban.trigger.builtin;
 
 import azkaban.trigger.ConditionChecker;
+import azkaban.utils.TimeUtils;
 import azkaban.utils.Utils;
 import java.util.Date;
 import java.util.HashMap;
@@ -88,7 +89,7 @@ public class BasicTimeChecker implements ConditionChecker {
     final boolean skipPastChecks =
         Boolean.valueOf((String) jsonObj.get("skipPastChecks"));
     final ReadablePeriod period =
-        Utils.parsePeriodString((String) jsonObj.get("period"));
+        TimeUtils.parsePeriodString((String) jsonObj.get("period"));
     final String id = (String) jsonObj.get("id");
     final String cronExpression = (String) jsonObj.get("cronExpression");
 
@@ -199,7 +200,7 @@ public class BasicTimeChecker implements ConditionChecker {
     jsonObj.put("nextCheckTime", String.valueOf(this.nextCheckTime));
     jsonObj.put("isRecurring", String.valueOf(this.isRecurring));
     jsonObj.put("skipPastChecks", String.valueOf(this.skipPastChecks));
-    jsonObj.put("period", Utils.createPeriodString(this.period));
+    jsonObj.put("period", TimeUtils.createPeriodString(this.period));
     jsonObj.put("id", this.id);
     jsonObj.put("cronExpression", this.cronExpression);
