+++ azkaban-common/src/test/java/azkaban/trigger/TriggerTest.java
@@ -23,7 +23,7 @@ import azkaban.trigger.builtin.BasicTimeChecker;
 import azkaban.trigger.builtin.ExecuteFlowAction;
 import azkaban.utils.JSONUtils;
 import azkaban.utils.Props;
-import azkaban.utils.Utils;
+import azkaban.utils.TimeUtils;
 import java.io.File;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -55,7 +55,7 @@ public class TriggerTest {
     final DateTime now = DateTime.now();
     final ConditionChecker checker1 =
         new BasicTimeChecker("timeChecker1", now.getMillis(), now.getZone(),
-            true, true, Utils.parsePeriodString("1h"), null);
+            true, true, TimeUtils.parsePeriodString("1h"), null);
     final Map<String, ConditionChecker> checkers1 =
         new HashMap<>();
     checkers1.put(checker1.getId(), checker1);
