+++ azkaban-web-server/src/main/java/azkaban/webapp/servlet/LoginAbstractAzkabanServlet.java
@@ -26,7 +26,6 @@ import azkaban.user.User;
 import azkaban.user.UserManager;
 import azkaban.user.UserManagerException;
 import azkaban.utils.StringUtils;
-import azkaban.utils.WebUtils;
 import azkaban.webapp.WebMetrics;
 import java.io.BufferedInputStream;
 import java.io.File;
@@ -220,9 +219,7 @@ public abstract class LoginAbstractAzkabanServlet extends
     headers.put(WebUtils.X_FORWARDED_FOR_HEADER,
         req.getHeader(WebUtils.X_FORWARDED_FOR_HEADER.toLowerCase()));
 
-    final WebUtils utils = new WebUtils();
-
-    return utils.getRealClientIpAddr(headers, req.getRemoteAddr());
+    return WebUtils.getRealClientIpAddr(headers, req.getRemoteAddr());
   }
 
   private Session getSessionFromRequest(final HttpServletRequest req)
@@ -328,9 +325,8 @@ public abstract class LoginAbstractAzkabanServlet extends
         // There are no valid sessions and temporary logins, no we either pass
         // back a message or redirect.
         if (isAjaxCall(req)) {
-          final String response =
-              createJsonResponse("error", "Invalid Session. Need to re-login",
-                  "login", null);
+          final String response = AbstractAzkabanServlet
+              .createJsonResponse("error", "Invalid Session. Need to re-login", "login", null);
           writeResponse(resp, response);
         } else {
           handleLogin(req, resp, "Enter username and password");
@@ -351,9 +347,9 @@ public abstract class LoginAbstractAzkabanServlet extends
    * distinction between the illegal request above and the following valid request: curl -X POST -d
    * "action=login&username=azkaban&password=azkaban" http://localhost:8081/
    *
-   * "password=" is searched for because it leverages the query syntax to determine that the user is
-   * passing the password as a parameter name. There is no other ajax call that has a parameter that
-   * includes the string "password" at the end which could throw false positives.
+   * "password=" is searched for because it leverages the query syntax to determine that the user
+   * is passing the password as a parameter name. There is no other ajax call that has a parameter
+   * that includes the string "password" at the end which could throw false positives.
    */
   private boolean isIllegalPostRequest(final HttpServletRequest req) {
     return (req.getQueryString() != null && req.getQueryString().contains("password="));
